requirements.txt.txt
Django==4.1
djangorestframework==3.14
Markdown==3.4.1
psycopg2==2.9.5